+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # 更改日志 | CHANGE LOG
 
 - 项目初始化 | Project initialization.
+- 添加附属仓库 `siyuan-plugin-jupyter-client` | Add sub-repository `siyuan-plugin-jupyter-client`.